fixed many compiler warnings and errors in x64 build in Visual Studio 2008
git-svn-id: file://localhost/home/tolik/svnrepo/trunk/opencv@1830 c5418bc3-7546-4aae-8bad-fa6f2f3e53e6

+++ src/cv/cvcalibinit.cpp
@@ -2157,7 +2157,7 @@ void drawChessboardCorners( Mat& image, Size patternSize,
 {
     CvMat _image = image;
     cvDrawChessboardCorners( &_image, patternSize, (CvPoint2D32f*)&corners[0],
-                             corners.size(), patternWasFound );
+                             (int)corners.size(), patternWasFound );
 }
 
 }
