+++ src/cxcore/_cxcore.h
@@ -233,9 +233,9 @@ binaryOpC1_( const Mat& srcmat1, const Mat& srcmat2, Mat& dstmat )
     const T1* src1 = (const T1*)srcmat1.data;
     const T2* src2 = (const T2*)srcmat2.data;
     DT* dst = (DT*)dstmat.data;
-    int step1 = srcmat1.step/sizeof(src1[0]);
-    int step2 = srcmat2.step/sizeof(src2[0]);
-    int step = dstmat.step/sizeof(dst[0]);
+    size_t step1 = srcmat1.step/sizeof(src1[0]);
+    size_t step2 = srcmat2.step/sizeof(src2[0]);
+    size_t step = dstmat.step/sizeof(dst[0]);
     Size size = getContinuousSize( srcmat1, srcmat2, dstmat, dstmat.channels() );
 
     if( size.width == 1 )
@@ -277,8 +277,8 @@ binarySOpCn_( const Mat& srcmat, Mat& dstmat, const Scalar& _scalar )
     typedef typename Op::rtype DT;
     const T* src0 = (const T*)srcmat.data;
     DT* dst0 = (DT*)dstmat.data;
-    int step1 = srcmat.step/sizeof(src0[0]);
-    int step = dstmat.step/sizeof(dst0[0]);
+    size_t step1 = srcmat.step/sizeof(src0[0]);
+    size_t step = dstmat.step/sizeof(dst0[0]);
     int cn = dstmat.channels();
     Size size = getContinuousSize( srcmat, dstmat, cn );
     WT scalar[12];
@@ -332,8 +332,8 @@ binarySOpC1_( const Mat& srcmat, Mat& dstmat, double _scalar )
     WT scalar = saturate_cast<WT>(_scalar);
     const T* src = (const T*)srcmat.data;
     DT* dst = (DT*)dstmat.data;
-    int step1 = srcmat.step/sizeof(src[0]);
-    int step = dstmat.step/sizeof(dst[0]);
+    size_t step1 = srcmat.step/sizeof(src[0]);
+    size_t step = dstmat.step/sizeof(dst[0]);
     Size size = srcmat.size();
     
     size.width *= srcmat.channels();
